AWS Lambda
Add the APP_SAVE_LOG environment variable to set if the application log must be written to a file
Add the aws-serverless-express package as a dependency. It acts as a wrapper to express in the lambda environment.
Add the lambda version driver (APP_VERSION_DRIVER environment variable)
Fix an erro in the applyCustomSaltToPassword method of the gl_user model
Allow the Vue config option publicPath to be set throught the ASSETS_DOMAIN environment variable (defaults to '/c/'). Assets can now be downloaded from S3 or a CDN, for example.
Add the lambda handler function